Comlpleted functions for every and some
The contains function was already existing and passing the given tests.
Wrote a function for every that used the reduce function, as suggested,
in order to get all the test cases to pass. Also wrote a function for
some that used the every function, as suggested, in order to get all the
test cases to pass.